+++ src/components/list/list.tsx
@@ -1,5 +1,5 @@
 import { IconSize } from '../icon/icon.types';
-import { ListItem, ListSeparator } from './list-item.types';
+import { ListItem, ListSeparator } from '../list-item/list-item.types';
 import { ListType } from './list.types';
 import { MDCList, MDCListActionEvent } from '@material/list';
 import { strings as listStrings } from '@material/list/constants';
