+++ src/components/picker/searcher.types.ts
@@ -1,4 +1,4 @@
-import { ListItem } from '../list/list-item.types';
+import { ListItem } from '../list-item/list-item.types';
 import { ListSeparator } from '../../global/shared-types/separator.types';
 
 /**
