+++ src/components/select/select.template.tsx
@@ -1,4 +1,4 @@
-import { ListItem, ListSeparator } from '../list/list-item.types';
+import { ListItem, ListSeparator } from '../list-item/list-item.types';
 import { Option } from '../select/option.types';
 import { FunctionalComponent, h } from '@stencil/core';
 import { isMultiple } from '../../util/multiple';
